old/requirements.txt
aio_pika
python-dotenv
aioredis